+++ src/main/java/org/springframework/http/server/reactive/RxNettyServerHttpResponse.java
@@ -95,12 +95,8 @@ public class RxNettyServerHttpResponse extends AbstractServerHttpResponse {
 				if (!httpCookie.getMaxAge().isNegative()) {
 					cookie.setMaxAge(httpCookie.getMaxAge().getSeconds());
 				}
-				if (httpCookie.getDomain().isPresent()) {
-					cookie.setDomain(httpCookie.getDomain().get());
-				}
-				if (httpCookie.getPath().isPresent()) {
-					cookie.setPath(httpCookie.getPath().get());
-				}
+				httpCookie.getDomain().ifPresent(cookie::setDomain);
+				httpCookie.getPath().ifPresent(cookie::setPath);
 				cookie.setSecure(httpCookie.isSecure());
 				cookie.setHttpOnly(httpCookie.isHttpOnly());
 				this.response.addCookie(cookie);
