+++ src/main/java/org/springframework/http/server/reactive/ServletServerHttpResponse.java
@@ -16,23 +16,17 @@
 
 package org.springframework.http.server.reactive;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.nio.charset.Charset;
 import java.util.List;
 import java.util.Map;
-import javax.servlet.ServletOutputStream;
-import javax.servlet.WriteListener;
+import java.util.function.Function;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletResponse;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.reactivestreams.Publisher;
-import org.reactivestreams.Subscriber;
-import org.reactivestreams.Subscription;
 import reactor.core.publisher.Mono;
-import reactor.core.util.BackpressureUtils;
 
 import org.springframework.core.io.buffer.DataBuffer;
 import org.springframework.core.io.buffer.DataBufferAllocator;
@@ -52,16 +46,16 @@ public class ServletServerHttpResponse extends AbstractServerHttpResponse {
 
 	private final HttpServletResponse response;
 
-	private final ResponseBodySubscriber responseBodySubscriber;
+	private final Function<Publisher<DataBuffer>, Mono<Void>> responseBodyWriter;
 
-	public ServletServerHttpResponse(ServletAsyncContextSynchronizer synchronizer,
-			int bufferSize, DataBufferAllocator allocator) throws IOException {
+	public ServletServerHttpResponse(HttpServletResponse response,
+			DataBufferAllocator allocator,
+			Function<Publisher<DataBuffer>, Mono<Void>> responseBodyWriter) {
 		super(allocator);
-		Assert.notNull(synchronizer, "'synchronizer' must not be null");
-
-		this.response = (HttpServletResponse) synchronizer.getResponse();
-		this.responseBodySubscriber =
-				new ResponseBodySubscriber(synchronizer, bufferSize);
+		Assert.notNull(response, "'response' must not be null");
+		Assert.notNull(responseBodyWriter, "'responseBodyWriter' must not be null");
+		this.response = response;
+		this.responseBodyWriter = responseBodyWriter;
 	}
 
 	public HttpServletResponse getServletResponse() {
@@ -75,8 +69,7 @@ public class ServletServerHttpResponse extends AbstractServerHttpResponse {
 
 	@Override
 	protected Mono<Void> setBodyInternal(Publisher<DataBuffer> publisher) {
-		return Mono.from((Publisher<Void>) subscriber -> publisher
-				.subscribe(this.responseBodySubscriber));
+		return this.responseBodyWriter.apply(publisher);
 	}
 
 	@Override
@@ -105,12 +98,8 @@ public class ServletServerHttpResponse extends AbstractServerHttpResponse {
 				if (!httpCookie.getMaxAge().isNegative()) {
 					cookie.setMaxAge((int) httpCookie.getMaxAge().getSeconds());
 				}
-				if (httpCookie.getDomain().isPresent()) {
-					cookie.setDomain(httpCookie.getDomain().get());
-				}
-				if (httpCookie.getPath().isPresent()) {
-					cookie.setPath(httpCookie.getPath().get());
-				}
+				httpCookie.getDomain().ifPresent(cookie::setDomain);
+				httpCookie.getPath().ifPresent(cookie::setPath);
 				cookie.setSecure(httpCookie.isSecure());
 				cookie.setHttpOnly(httpCookie.isHttpOnly());
 				this.response.addCookie(cookie);
@@ -118,140 +107,4 @@ public class ServletServerHttpResponse extends AbstractServerHttpResponse {
 		}
 	}
 
-	private static class ResponseBodySubscriber implements Subscriber<DataBuffer> {
-
-		private final ResponseBodyWriteListener writeListener =
-				new ResponseBodyWriteListener();
-
-		private final ServletAsyncContextSynchronizer synchronizer;
-
-		private final int bufferSize;
-
-		private volatile DataBuffer dataBuffer;
-
-		private volatile boolean completed = false;
-
-		private Subscription subscription;
-
-		public ResponseBodySubscriber(ServletAsyncContextSynchronizer synchronizer,
-				int bufferSize) throws IOException {
-			this.synchronizer = synchronizer;
-			this.bufferSize = bufferSize;
-			synchronizer.getResponse().getOutputStream().setWriteListener(writeListener);
-		}
-
-		@Override
-		public void onSubscribe(Subscription subscription) {
-			logger.trace("onSubscribe. Subscription: " + subscription);
-			if (BackpressureUtils.validate(this.subscription, subscription)) {
-				this.subscription = subscription;
-				this.subscription.request(1);
-			}
-		}
-
-		@Override
-		public void onNext(DataBuffer dataBuffer) {
-			Assert.state(this.dataBuffer == null);
-
-			logger.trace("onNext. buffer: " + dataBuffer);
-
-			this.dataBuffer = dataBuffer;
-			try {
-				this.writeListener.onWritePossible();
-			}
-			catch (IOException e) {
-				onError(e);
-			}
-		}
-
-		@Override
-		public void onError(Throwable t) {
-			logger.error("onError", t);
-			HttpServletResponse response =
-					(HttpServletResponse) this.synchronizer.getResponse();
-			response.setStatus(HttpStatus.INTERNAL_SERVER_ERROR.value());
-			this.synchronizer.complete();
-
-		}
-
-		@Override
-		public void onComplete() {
-			logger.trace("onComplete. buffer: " + this.dataBuffer);
-
-			this.completed = true;
-
-			if (this.dataBuffer != null) {
-				try {
-					this.writeListener.onWritePossible();
-				}
-				catch (IOException ex) {
-					onError(ex);
-				}
-			}
-
-			if (this.dataBuffer == null) {
-				this.synchronizer.writeComplete();
-			}
-		}
-
-		private class ResponseBodyWriteListener implements WriteListener {
-
-			@Override
-			public void onWritePossible() throws IOException {
-				logger.trace("onWritePossible");
-				ServletOutputStream output = synchronizer.getResponse().getOutputStream();
-
-				boolean ready = output.isReady();
-				logger.trace("ready: " + ready + " buffer: " + dataBuffer);
-
-				if (ready) {
-					if (dataBuffer != null) {
-
-						int total = dataBuffer.readableByteCount();
-						int written = writeDataBuffer();
-
-						logger.trace("written: " + written + " total: " + total);
-						if (written == total) {
-							releaseBuffer();
-							if (!completed) {
-								subscription.request(1);
-							}
-							else {
-								synchronizer.writeComplete();
-							}
-						}
-					}
-					else if (subscription != null) {
-						subscription.request(1);
-					}
-				}
-			}
-
-			private int writeDataBuffer() throws IOException {
-				InputStream input = dataBuffer.asInputStream();
-				ServletOutputStream output = synchronizer.getResponse().getOutputStream();
-
-				int bytesWritten = 0;
-				byte[] buffer = new byte[bufferSize];
-				int bytesRead = -1;
-
-				while (output.isReady() && (bytesRead = input.read(buffer)) != -1) {
-					output.write(buffer, 0, bytesRead);
-					bytesWritten += bytesRead;
-				}
-
-				return bytesWritten;
-			}
-
-			private void releaseBuffer() {
-				// TODO: call PooledDataBuffer.release() when we it is introduced
-				dataBuffer = null;
-			}
-
-			@Override
-			public void onError(Throwable ex) {
-				logger.error("ResponseBodyWriteListener error", ex);
-			}
-		}
-	}
 }
